src/veille_pqc/config_loader.py
from __future__ import annotations

from pathlib import Path

import yaml

from veille_pqc.models import AppConfig


def load_config(path: str | Path) -> AppConfig:
    config_path = Path(path)
    with config_path.open("r", encoding="utf-8") as f:
        data = yaml.safe_load(f)
    return AppConfig.model_validate(data)
